Refactor: Rename node assets module
